requirements.txt
Django>=5.0
djangorestframework
psycopg2-binary
boto3
django-storages
gunicorn